requirements.txt
click==8.0.1
Jinja2==3.0.1
PyYAML==5.4.1
